requirements.txt
certifi==2018.1.18
future==0.16.0
parsedatetime==2.4
python-telegram-bot==10.0.1
